+++ src/Spreads.Collections/Common/Types.fs
@@ -47,3 +47,10 @@ type ValueTuple<'V1,'V2,'V3> =
     new(v1 : 'V1, v2 : 'V2, v3 : 'V3) = {Value1 = v1; Value2 = v2; Value3 = v3}
     new(tuple2:ValueTuple<'V1,'V2>, v3 : 'V3) = {Value1 = tuple2.Value1; Value2 = tuple2.Value2; Value3 = v3}
   end
+
+
+type DummyDisposable private() =
+  static let instance = new DummyDisposable()
+  interface IDisposable with
+    member x.Dispose() = ()
+  static member Instance with get() = instance :> IDisposable